refactor(params): ParamControl 型別移出 hook（解除 state → hooks 反向 import）
`SliderConfig` / `ToggleConfig` / `SelectConfig` / `ParamControl` 從
`hooks/useTransportParams.ts` 搬到 `state/layerParamsControls.ts` ——
後者是這組型別唯一的生產者（`buildParamControls`），原本卻得反過來
`import type ... from "../hooks/…"`，`hooks → state` 相依是倒的。

搬完相依單向：`data/layerParamsSpec → state/* → hooks/* ＋ components/*`。

⚠️ `SliderConfig.type?: "slider"` 的**選填**性質逐字不動（P3-1 記過：
補一個 `type: "slider"` 會讓黃金快照 params section 立刻紅）。

import 路徑同步 4 處：IconRailSidebar / LayerSidebar / layerParamsControls.test /
useTransportParams 自己（改成正向 import）。

驗收：tsc 0 error；vitest 43 檔 567 passed / 1 skipped；
黃金快照 sha256 07972fce… 未變。
